+++ paddlenlp/transformers/utils.py
@@ -689,43 +689,6 @@ def get_checkpoint_shard_files(
     return cached_filenames, sharded_metadata
 
 
-def get_optimizer_shard_files(optimizer_path, index_filename):
-    """
-    For a given model:
-    - download and cache all the shards of a sharded checkpoint if `pretrained_model_name_or_path` is a model ID on the
-      Hub
-    - returns the list of paths to all the shards, as well as some metadata.
-    For the description of each arg, see [`PretrainedModel.from_pretrained`]. `index_filename` is the full path to the
-    index (downloaded and cached if `pretrained_model_name_or_path` is a model ID on the Hub).
-    """
-
-    import json
-
-    if not os.path.isfile(index_filename):
-        raise ValueError(f"Can't find a optimizer index ({index_filename}) in {optimizer_path}.")
-
-    with open(index_filename, "r") as f:
-        index = json.loads(f.read())
-
-    shard_filenames = sorted(set(index["weight_map"].values()))
-    sharded_metadata = index["metadata"]
-    sharded_metadata["all_optimizer_keys"] = list(index["weight_map"].keys())
-    sharded_metadata["weight_map"] = index["weight_map"].copy()
-    sharded_metadata["master_weights"] = index.get("master_weights", False)
-
-    file_map = {file: set() for file in shard_filenames}
-    for weight, file in index["weight_map"].items():
-        file_map[file].add(weight)
-
-    sharded_metadata["file_map"] = file_map
-
-    # First, let's deal with local folder.
-    # TODO: if optimizer_path is a folder, we should check if the optimizer is already cached or not.
-    if os.path.isdir(optimizer_path):
-        shard_filenames = [os.path.join(optimizer_path, f) for f in shard_filenames]
-        return shard_filenames, sharded_metadata
-
-
 def is_safetensors_available():
     return importlib.util.find_spec("safetensors") is not None
